[VIEWS] restructure
