+++ app/webapp/utils/iiif/gen_html.py
@@ -28,17 +28,17 @@ def regions_btn(obj, action="view"):
         icon = get_icon("eye")
         # The link redirects to Mirador with no regions (Digitization) or automatic regions (Regions)
         link = f"{SAS_APP_URL}/indexView.html?iiif-content={obj.gen_manifest_url(version=MANIFEST_V1)}"
-    elif action == "edit":
-        icon = get_icon("pen-to-square")
-        # The link redirects to the edit regions page (show_regions() view)
-        link = f"{APP_URL}/{APP_NAME}/{obj.get_ref()}/show/"
+    # elif action == "edit":
+    #     icon = get_icon("pen-to-square")
+    #     # The link redirects to the edit regions page (show_regions() view DELETED)
+    #     link = f"{APP_URL}/{APP_NAME}/{obj.get_ref()}/show/"
     elif action == "final":
         icon = get_icon("check")
         # The link redirects to Mirador with corrected regions (Regions)
         link = f"{SAS_APP_URL}/indexView.html?iiif-content={obj.gen_manifest_url(version=MANIFEST_V2)}"
-    elif action == "similarity":
-        icon = get_icon("code-compare")
-        link = f"{APP_URL}/{APP_NAME}/{obj.get_ref()}/show-similarity"
+    # elif action == "similarity":
+    #     icon = get_icon("code-compare")
+    #     link = f"{APP_URL}/{APP_NAME}/{obj.get_ref()}/show-similarity"
     elif action == "regions":
         icon = get_icon("eye")
         link = f"{APP_URL}/{APP_NAME}/{obj.get_ref()}/show-all-regions"
@@ -90,18 +90,21 @@ def gen_btn(obj, action="view"):
     # if action == "regions" or action == "vectors":
     #     return mark_safe(f"<br>{regions_btn(obj, action)}")
 
-    if action == "auto-view":
-        digit_id = obj.id if cls(obj) == Digitization else obj.get_digit().id
-        download_url = f"{APP_URL}/{APP_NAME}/iiif/digit-regions/{digit_id}"
-        regions_type = "TXT"
-        version = None if cls(obj) == Digitization else MANIFEST_V1
-        is_regions = obj.has_regions() if cls(obj) == Digitization else True
-    else:
-        download_url = f"{SAS_APP_URL}/search-api/{obj.get_ref()}/search/"
-        regions_type = "JSON"
-        version = MANIFEST_V2
+    # if action == "auto-view":
+    #     digit_id = obj.id if cls(obj) == Digitization else obj.get_digit().id
+    #     download_url = f"{APP_URL}/{APP_NAME}/iiif/digit-regions/{digit_id}"
+    #     regions_type = "TXT"
+    #     version = None if cls(obj) == Digitization else MANIFEST_V1
+    #     is_regions = obj.has_regions() if cls(obj) == Digitization else True
+    # else:
+    #     download_url = f"{SAS_APP_URL}/search-api/{obj.get_ref()}/search/"
+    #     regions_type = "JSON"
+    #     version = MANIFEST_V2
     # else:
     #     return mark_safe("NOT SUPPOSED TO OCCUR")
+    download_url = f"{SAS_APP_URL}/search-api/{obj.get_ref()}/search/"
+    regions_type = "JSON"
+    version = MANIFEST_V2
 
     download = (
         f'<a href="{download_url}" target="_blank">{get_icon("download")} {get_action("download")} ({regions_type})</a>'
